Disable automatic snapshotting
Summary:
We want better control over the snapshotting, so we'll do it
only in Lisp.

The context of this was that Reuben pinged me about a long running API
call, and we saw in Preset that a snapshot was happening at exactly
the same time, and snapshot times have been trending upwards.

Differential Revision: https://phabricator.tdrhq.com/D9298

GitOrigin-RevId: 566aef7c436485261305e53dd967f385d2aa008c